refactor: Shift classes and add javadocs
