+++ src/main/java/com/onarandombox/MultiverseCore/commands/CreateCommand.java
@@ -22,7 +22,7 @@ import com.onarandombox.MultiverseCore.commandtools.flags.CommandFlag;
 import com.onarandombox.MultiverseCore.commandtools.flags.CommandFlagGroup;
 import com.onarandombox.MultiverseCore.commandtools.flags.CommandValueFlag;
 import com.onarandombox.MultiverseCore.commandtools.flags.ParsedCommandFlags;
-import com.onarandombox.MultiverseCore.locale.MVCorei18n;
+import com.onarandombox.MultiverseCore.utils.MVCorei18n;
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.WorldType;
