--- src/main/java/com/onarandombox/MultiverseCore/locale/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains all the locale reference for Multiverse-Core.
- */
-package com.onarandombox.MultiverseCore.locale;
