+++ src/main/java/com/onarandombox/MultiverseCore/utils/MVCorei18n.java
@@ -1,4 +1,4 @@
-package com.onarandombox.MultiverseCore.locale;
+package com.onarandombox.MultiverseCore.utils;
 
 import co.aikar.locales.MessageKey;
 import co.aikar.locales.MessageKeyProvider;
