+++ src/main/java/com/onarandombox/MultiverseCore/utils/file/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains all the classes that deal with files on disk.
+ */
+package com.onarandombox.MultiverseCore.utils.file;
